requirements.txt
Django==4.0.3
pysaml2==7.1.2
pydantic==1.9.0